🧹 Refactor duplicated coordinate bounding box fallback logic
Extracted `fallbackBox` and `fallbackPathForCode` into a new `apps/inc/geo_helpers.php` file.
Replaced manual string concatenation in `fallbackBox` with `json_encode` for safer, more robust output.
Updated `apps/inc/geo_ajax.php` and `apps/inc/reverse_lookup.php` to use the shared helper.
